packages/bot-telegram/pyproject.toml
[project]
name = "bot-telegram"
version = "0.1.0"
description = "Telegram бот - ai-бот городской помощник (для Санкт-Петербурга) в Telegram-мессенджере"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.22.0",
    "python-dotenv>=1.2.1",
    "structlog>=25.5.0",
    "pydantic>=2.0,<2.12",
    "rich>=14.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bot_telegram"]

[tool.uv.sources]
langgraph-client = { workspace = true }
